requirements.txt
aiohttp==3.6.2
aioredis==1.3.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
hiredis==1.0.1
idna==2.8
multidict==4.7.4
PyYAML==5.3
requests==2.22.0
urllib3==1.25.8
yarl==1.4.2
